requirements.txt
Flask>=1.0,<=1.1.2
pandas==1.2.4
numpy>=1.16.5
azure-storage-blob==12.8.1